feat: update VenuesController to allow access for both ADMIN and PROFESSOR roles; enhance VenuesService to prevent deletion of venues with associated advisory dates

+++ src/venues/venues.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Venue } from './entities/venue.entity';
@@ -7,12 +7,15 @@ import { UpdateVenueDto } from './dto/update-venue.dto';
 import { PaginatedResult } from 'src/common/interfaces/paginated-result.interface';
 import { VenueDto } from './dto/venue.dto';
 import { VenueQueryDto } from './dto/venue-query.dto';
+import { AdvisoryDate } from 'src/advisory-dates/entities/advisory-date.entity';
 
 @Injectable()
 export class VenuesService {
   constructor(
     @InjectRepository(Venue)
     private readonly repo: Repository<Venue>,
+    @InjectRepository(AdvisoryDate)
+    private readonly advisoryDateRepo: Repository<AdvisoryDate>,
   ) {}
 
   create(dto: CreateVenueDto) {
@@ -72,7 +75,23 @@ export class VenuesService {
     return this.findOne(id);
   }
 
-  remove(id: number) {
+  async remove(id: number) {
+    // Check if the venue exists
+    const venue = await this.findOne(id);
+    if (!venue) {
+      throw new Error('Venue not found');
+    }
+
+    // Check if exist any advisory dates associated with this venue
+    const advisoryDates = await this.advisoryDateRepo.find({
+      where: { venue: { venue_id: id } },
+    });
+    if (advisoryDates.length > 0) {
+      throw new BadRequestException(
+        'Cannot delete venue with associated advisory dates',
+      );
+    }
+
     return this.repo.delete(id);
   }
 }
